auto-chain/requirements.txt
python-dotenv==1.0.0
langchain==0.0.273
openai==0.27.8
pypdf==3.15.0
matplotlib==3.5.1
tiktoken==0.4.0
faiss-cpu==1.7.4
st-annotated-text==4.0.0
unstructured==0.9.0
pdf2image==1.16.3
pdfminer.six==20221105
pytesseract==0.3.10
docx2txt==0.8
transformers==4.31.0
torch==2.0.1
firebase-admin==6.0.1
streamlit-tags==1.2.8
extra-streamlit-components==0.1.60
